backend/pyproject.toml
# ============================================
# Python 项目配置文件 (pyproject.toml)
#
# 现代 Python 项目的标准配置文件，替代了 setup.py、setup.cfg 等
# 使用 uv 作为包管理工具
# ============================================

# ============================================
# 项目基本信息
# ============================================
[project]
name = "app"
version = "0.1.0"
description = ""
# Python 版本要求：3.10 及以上，4.0 以下
requires-python = ">=3.10,<4.0"

# ============================================
# 生产环境依赖 (Dependencies)
# ============================================
dependencies = [
    # FastAPI - 高性能 Web 框架
    # [standard] 包含 uvicorn 服务器和常用依赖
    "fastapi[standard]<1.0.0,>=0.114.2",
    
    # 处理 multipart/form-data 请求（文件上传等）
    "python-multipart<1.0.0,>=0.0.7",
    
    # 邮箱地址验证
    "email-validator<3.0.0.0,>=2.1.0.post1",
    
    # 重试机制 - 用于数据库连接等场景
    "tenacity<9.0.0,>=8.2.3",
    
    # Pydantic v2 - 数据验证和序列化
    "pydantic>2.0",
    
    # 发送邮件
    "emails<1.0,>=0.6",
    
    # Jinja2 模板引擎 - 用于渲染 HTML 邮件模板
    "jinja2<4.0.0,>=3.1.4",
    
    # Alembic - 数据库迁移工具
    "alembic<2.0.0,>=1.12.1",
    
    # HTTP 客户端 - 用于测试和外部 API 调用
    "httpx<1.0.0,>=0.25.1",
    
    # MySQL 数据库驱动（原为 PostgreSQL 的 psycopg）
    "pymysql<2.0.0,>=1.1.0",
    
    # 加密库 - MySQL 8.0 认证需要
    "cryptography>=42.0.0",
    
    # SQLModel - SQLAlchemy + Pydantic 的 ORM
    "sqlmodel<1.0.0,>=0.0.21",
    
    # Pydantic Settings - 配置管理
    "pydantic-settings<3.0.0,>=2.2.1",
    
    # Sentry - 错误追踪和性能监控
    "sentry-sdk[fastapi]<2.0.0,>=1.40.6",
    
    # PyJWT - JWT Token 生成和验证
    "pyjwt<3.0.0,>=2.8.0",
    
    # 密码哈希 - 支持 Argon2 和 BCrypt 算法
    "pwdlib[argon2,bcrypt]>=0.3.0",
    
    # APScheduler - 定时任务调度
    "apscheduler>=3.10.0,<4.0.0",
    
    # Pandas - 数据处理和 CSV/Excel 解析
    "pandas>=2.0.0",
    
    # OpenPyXL - Excel 文件解析
    "openpyxl>=3.1.0",
]

# ============================================
# 开发环境依赖 (Dev Dependencies)
# ============================================
[dependency-groups]
dev = [
    # Pytest - 测试框架
    "pytest<8.0.0,>=7.4.3",
    
    # MyPy - 静态类型检查
    "mypy<2.0.0,>=1.8.0",
    
    # Ruff - 极速 Python 代码检查工具（替代 flake8、black、isort）
    "ruff<1.0.0,>=0.2.2",
    
    # Prek - 预提交钩子工具（替代 pre-commit）
    "prek>=0.2.24,<1.0.0",
    
    # Coverage - 测试覆盖率统计
    "coverage<8.0.0,>=7.4.3",
]

# ============================================
# 构建系统配置
# ============================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ============================================
# MyPy 类型检查配置
# ============================================
[tool.mypy]
# 严格模式 - 启用所有类型检查
strict = true
# 排除的目录
exclude = ["venv", ".venv", "alembic"]

# ============================================
# Ruff 代码检查配置
# ============================================
[tool.ruff]
# 目标 Python 版本
target-version = "py310"
# 排除的目录
exclude = ["alembic"]

[tool.ruff.lint]
# 启用的规则集
select = [
    "E",        # pycodestyle 错误
    "W",        # pycodestyle 警告
    "F",        # Pyflakes
    "I",        # isort（导入排序）
    "B",        # flake8-bugbear（常见 bug 检查）
    "C4",       # flake8-comprehensions（推导式优化）
    "UP",       # pyupgrade（Python 升级建议）
    "ARG001",   # 未使用的函数参数
    "T201",     # 禁止 print 语句（生产代码应使用 logging）
]
# 忽略的规则
ignore = [
    "E501",     # 行过长（由编辑器自动处理）
    "B008",     # 允许在参数默认值中调用函数
    "W191",     # 缩进包含制表符
    "B904",     # 允许直接 raise HTTPException 而不使用 from e
]

[tool.ruff.lint.pyupgrade]
# 保留运行时类型注解，即使文件导入了 from __future__ import annotations
keep-runtime-typing = true

# ============================================
# 测试覆盖率配置
# ============================================
[tool.coverage.run]
# 统计覆盖率的源代码目录
source = ["app"]
# 动态上下文 - 按测试函数统计
dynamic_context = "test_function"

[tool.coverage.report]
# 显示未覆盖的行号
show_missing = true
# 按覆盖率排序（高到低）
sort = "-Cover"

[tool.coverage.html]
# HTML 报告显示上下文
show_contexts = true
